WoT restructure: nest overlay class hierarchy into overlay/regions/
Mirrors table/regions/: the abstract Overlay base (_base.ts) + the 5 concrete overlay
subclasses (top/bottom/inlineStart + 2 corners) move into overlay/regions/. overlay/ top
keeps the coordinator (overlays.ts), constants, barrel (index.ts), resizeMonitor, strategies/.
Import paths shifted one level; all _base importers (barrel, overlays, stickyScrollStrategy,
core/clone, core/core, facade/core, wire, types, plugins/mergeCells) repointed to
overlay/regions/_base. Pure relocation. Gated green: test:types + eslint + test:walkontable 749/0/2.
