+++ handsontable/src/3rdparty/walkontable/src/core/clone.ts
@@ -2,7 +2,7 @@ import Event, { createEventDeps } from '../event';
 import { createTableDeps } from '../table/baseTable';
 import CoreAbstract from './_base';
 import type Settings from '../settings';
-import type { CloneDeps } from '../overlay/_base';
+import type { CloneDeps } from '../overlay/regions/_base';
 
 /**
  * @class Walkontable
