+++ handsontable/src/3rdparty/walkontable/src/overlay/index.ts
@@ -1,7 +1,7 @@
-export { BottomInlineStartCornerOverlay } from './bottomInlineStartCornerOverlay';
-export { BottomOverlay } from './bottomOverlay';
-export { InlineStartOverlay } from './inlineStartOverlay';
-export { Overlay } from './_base';
-export { TopInlineStartCornerOverlay } from './topInlineStartCornerOverlay';
-export { TopOverlay } from './topOverlay';
+export { BottomInlineStartCornerOverlay } from './regions/bottomInlineStartCornerOverlay';
+export { BottomOverlay } from './regions/bottomOverlay';
+export { InlineStartOverlay } from './regions/inlineStartOverlay';
+export { Overlay } from './regions/_base';
+export { TopInlineStartCornerOverlay } from './regions/topInlineStartCornerOverlay';
+export { TopOverlay } from './regions/topOverlay';
 export * from './constants';
