+++ handsontable/src/3rdparty/walkontable/src/overlay/regions/bottomInlineStartCornerOverlay.ts
@@ -1,12 +1,12 @@
-import type { TableDeps } from '../table/baseTable';
+import type { TableDeps } from '../../table/baseTable';
 import {
   resetCssTransform
-} from '../../../../helpers/dom/element';
-import BottomInlineStartCornerOverlayTable from '../table/regions/bottomInlineStartCornerTable';
+} from '../../../../../helpers/dom/element';
+import BottomInlineStartCornerOverlayTable from '../../table/regions/bottomInlineStartCornerTable';
 import { Overlay, type OverlayDeps } from './_base';
 import {
   CLONE_BOTTOM_INLINE_START_CORNER,
-} from './constants';
+} from '../constants';
 
 /**
  * @class BottomInlineStartCornerOverlay
